refactor(packages): move query and script packages to shell
- relocate `query` and `script` packages under the new `shell` package
- update all import statements to reflect the new package structure
- remove old `package-info.java` files and add a new one for `shell`
