+++ src/main/java/io/github/eggy03/ferrumx/windows/entity/storage/Win32DiskDriveToDiskPartition.java
@@ -9,7 +9,7 @@ import com.google.errorprone.annotations.Immutable;
 import com.google.gson.GsonBuilder;
 import com.google.gson.annotations.SerializedName;
 import io.github.eggy03.ferrumx.windows.annotation.WmiClass;
-import io.github.eggy03.ferrumx.windows.query.Cimv2;
+import io.github.eggy03.ferrumx.windows.shell.query.Cimv2;
 import lombok.Builder;
 import lombok.Value;
 import org.jetbrains.annotations.NotNull;
