--- src/main/java/io/github/eggy03/ferrumx/windows/query/package-info.java
@@ -1,15 +1,0 @@
-/**
- * Contains predefined PowerShell queries for Windows WMI and CIM classes,
- * organized by their respective namespaces. The queries are in WQL format.
- * <p>
- * These queries are executed by classes in the {@link io.github.eggy03.ferrumx.windows.service} package
- * to retrieve system information from Windows Management Instrumentation (WMI).
- * </p>
- * <p>
- * <strong>Note:</strong> The contents of this package are intended for internal use within the library.
- * </p>
- *
- * @author Sayan Bhattacharjee (Egg-03/Eggy)
- * @since 4.1.0
- */
-package io.github.eggy03.ferrumx.windows.query;
