--- src/main/java/io/github/eggy03/ferrumx/windows/script/package-info.java
@@ -1,11 +1,0 @@
-/**
- * <p>
- * Contains predefined PowerShell script paths and a loader for loading the scripts as streams.
- * </p>
- * <strong>Note:</strong> The contents of this package are intended for internal use within the library.
- * </p>
- *
- * @author Sayan Bhattacharjee (Egg-03/Eggy)
- * @since 4.1.0
- */
-package io.github.eggy03.ferrumx.windows.script;
