+++ src/main/java/io/github/eggy03/ferrumx/windows/service/compounded/HardwareIdService.java
@@ -9,9 +9,9 @@ import com.profesorfalken.jpowershell.PowerShell;
 import com.profesorfalken.jpowershell.PowerShellResponse;
 import io.github.eggy03.ferrumx.windows.entity.compounded.HardwareId;
 import io.github.eggy03.ferrumx.windows.mapping.compounded.HardwareIdMapper;
-import io.github.eggy03.ferrumx.windows.script.ScriptEnum;
-import io.github.eggy03.ferrumx.windows.script.ScriptLoader;
 import io.github.eggy03.ferrumx.windows.service.OptionalCommonServiceInterface;
+import io.github.eggy03.ferrumx.windows.shell.script.ScriptEnum;
+import io.github.eggy03.ferrumx.windows.shell.script.ScriptLoader;
 import io.github.eggy03.ferrumx.windows.utility.TerminalUtility;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
