+++ src/main/java/io/github/eggy03/ferrumx/windows/service/compounded/MsftNetAdapterToIpAndDnsAndProfileService.java
@@ -9,13 +9,13 @@ import com.profesorfalken.jpowershell.PowerShell;
 import com.profesorfalken.jpowershell.PowerShellResponse;
 import io.github.eggy03.ferrumx.windows.entity.compounded.MsftNetAdapterToIpAndDnsAndProfile;
 import io.github.eggy03.ferrumx.windows.mapping.compounded.MsftNetAdapterToIpAndDnsAndProfileMapper;
-import io.github.eggy03.ferrumx.windows.script.ScriptEnum;
-import io.github.eggy03.ferrumx.windows.script.ScriptLoader;
 import io.github.eggy03.ferrumx.windows.service.CommonServiceInterface;
 import io.github.eggy03.ferrumx.windows.service.network.MsftDnsClientServerAddressService;
 import io.github.eggy03.ferrumx.windows.service.network.MsftNetAdapterService;
 import io.github.eggy03.ferrumx.windows.service.network.MsftNetConnectionProfileService;
 import io.github.eggy03.ferrumx.windows.service.network.MsftNetIpAddressService;
+import io.github.eggy03.ferrumx.windows.shell.script.ScriptEnum;
+import io.github.eggy03.ferrumx.windows.shell.script.ScriptLoader;
 import io.github.eggy03.ferrumx.windows.utility.TerminalUtility;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
