+++ src/main/java/io/github/eggy03/ferrumx/windows/service/compounded/Win32DiskPartitionToLogicalDiskService.java
@@ -10,14 +10,14 @@ import com.profesorfalken.jpowershell.PowerShellResponse;
 import io.github.eggy03.ferrumx.windows.entity.compounded.Win32DiskDriveToPartitionAndLogicalDisk;
 import io.github.eggy03.ferrumx.windows.entity.compounded.Win32DiskPartitionToLogicalDisk;
 import io.github.eggy03.ferrumx.windows.mapping.compounded.Win32DiskPartitionToLogicalDiskMapper;
-import io.github.eggy03.ferrumx.windows.script.ScriptEnum;
-import io.github.eggy03.ferrumx.windows.script.ScriptLoader;
 import io.github.eggy03.ferrumx.windows.service.CommonServiceInterface;
 import io.github.eggy03.ferrumx.windows.service.storage.Win32DiskDriveService;
 import io.github.eggy03.ferrumx.windows.service.storage.Win32DiskDriveToDiskPartitionService;
 import io.github.eggy03.ferrumx.windows.service.storage.Win32DiskPartitionService;
 import io.github.eggy03.ferrumx.windows.service.storage.Win32LogicalDiskService;
 import io.github.eggy03.ferrumx.windows.service.storage.Win32LogicalDiskToPartitionService;
+import io.github.eggy03.ferrumx.windows.shell.script.ScriptEnum;
+import io.github.eggy03.ferrumx.windows.shell.script.ScriptLoader;
 import io.github.eggy03.ferrumx.windows.utility.TerminalUtility;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
