+++ src/main/java/io/github/eggy03/ferrumx/windows/service/compounded/Win32ProcessorToCacheMemoryService.java
@@ -9,12 +9,12 @@ import com.profesorfalken.jpowershell.PowerShell;
 import com.profesorfalken.jpowershell.PowerShellResponse;
 import io.github.eggy03.ferrumx.windows.entity.compounded.Win32ProcessorToCacheMemory;
 import io.github.eggy03.ferrumx.windows.mapping.compounded.Win32ProcessorToCacheMemoryMapper;
-import io.github.eggy03.ferrumx.windows.script.ScriptEnum;
-import io.github.eggy03.ferrumx.windows.script.ScriptLoader;
 import io.github.eggy03.ferrumx.windows.service.CommonServiceInterface;
 import io.github.eggy03.ferrumx.windows.service.processor.Win32AssociatedProcessorMemoryService;
 import io.github.eggy03.ferrumx.windows.service.processor.Win32CacheMemoryService;
 import io.github.eggy03.ferrumx.windows.service.processor.Win32ProcessorService;
+import io.github.eggy03.ferrumx.windows.shell.script.ScriptEnum;
+import io.github.eggy03.ferrumx.windows.shell.script.ScriptLoader;
 import io.github.eggy03.ferrumx.windows.utility.TerminalUtility;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
