+++ src/main/java/io/github/eggy03/ferrumx/windows/service/network/Win32NetworkAdapterSettingService.java
@@ -11,8 +11,8 @@ import io.github.eggy03.ferrumx.windows.entity.network.Win32NetworkAdapter;
 import io.github.eggy03.ferrumx.windows.entity.network.Win32NetworkAdapterConfiguration;
 import io.github.eggy03.ferrumx.windows.entity.network.Win32NetworkAdapterSetting;
 import io.github.eggy03.ferrumx.windows.mapping.network.Win32NetworkAdapterSettingMapper;
-import io.github.eggy03.ferrumx.windows.query.Cimv2;
 import io.github.eggy03.ferrumx.windows.service.CommonServiceInterface;
+import io.github.eggy03.ferrumx.windows.shell.query.Cimv2;
 import io.github.eggy03.ferrumx.windows.utility.TerminalUtility;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
