+++ src/main/java/io/github/eggy03/ferrumx/windows/service/peripheral/Win32PrinterService.java
@@ -9,8 +9,8 @@ import com.profesorfalken.jpowershell.PowerShell;
 import com.profesorfalken.jpowershell.PowerShellResponse;
 import io.github.eggy03.ferrumx.windows.entity.peripheral.Win32Printer;
 import io.github.eggy03.ferrumx.windows.mapping.peripheral.Win32PrinterMapper;
-import io.github.eggy03.ferrumx.windows.query.Cimv2;
 import io.github.eggy03.ferrumx.windows.service.CommonServiceInterface;
+import io.github.eggy03.ferrumx.windows.shell.query.Cimv2;
 import io.github.eggy03.ferrumx.windows.utility.TerminalUtility;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
