+++ src/main/java/io/github/eggy03/ferrumx/windows/shell/query/Cimv2.java
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  * See the LICENSE file in the project root for more information.
  */
-package io.github.eggy03.ferrumx.windows.query;
+package io.github.eggy03.ferrumx.windows.shell.query;
 
 import com.profesorfalken.jpowershell.PowerShell;
 import io.github.eggy03.ferrumx.windows.entity.display.Win32DesktopMonitor;
@@ -28,7 +28,6 @@ import io.github.eggy03.ferrumx.windows.entity.system.Win32OperatingSystem;
 import io.github.eggy03.ferrumx.windows.entity.system.Win32PnPEntity;
 import io.github.eggy03.ferrumx.windows.entity.system.Win32Process;
 import io.github.eggy03.ferrumx.windows.entity.user.Win32UserAccount;
-import io.github.eggy03.ferrumx.windows.query.utility.QueryUtility;
 import lombok.Getter;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
